requirements.txt
pyserial
pyaml
python-dotenv
bitdotio[psycopg2-binary]